Subtitle Styling (#193)
* added vlc subtitle options

* split swift code to multiple files

* updated swift codes

* Bump to 5.0.5

* added subtitle enums

update example code for subtitle styling

* Update flutter_vlc_player_platform_interface/lib/src/enums/subtitles/vlc_subtitle_text_direction.dart


* Update flutter_vlc_player_platform_interface/lib/src/enums/subtitles/vlc_subtitle_color.dart


* Update flutter_vlc_player_platform_interface/lib/src/utils/options/vlc_subtitle_options.dart


* updated subtitle color styling implementation

* updated subtitle thickness option
