+++ flutter_vlc_player/example/lib/single_tab.dart
@@ -92,6 +92,14 @@ class _SingleTabState extends State<SingleTab> {
             advanced: VlcAdvancedOptions([
               VlcAdvancedOptions.networkCaching(2000),
             ]),
+            subtitle: VlcSubtitleOptions([
+              VlcSubtitleOptions.boldStyle(true),
+              VlcSubtitleOptions.fontSize(30),
+              VlcSubtitleOptions.outlineColor(VlcSubtitleColor.yellow),
+              VlcSubtitleOptions.outlineThickness(VlcSubtitleThickness.normal),
+              // works only on externally added subtitles
+              VlcSubtitleOptions.color(VlcSubtitleColor.navy),
+            ]),
             rtp: VlcRtpOptions([
               VlcRtpOptions.rtpOverRtsp(true),
             ]),
