+++ flutter_vlc_player/pubspec.yaml
@@ -1,6 +1,6 @@
 name: flutter_vlc_player
 description: A VLC-powered alternative to Flutter's video_player. Supports multiple players on one screen.
-version: 5.0.4
+version: 5.0.5
 homepage: https://github.com/solid-software/flutter_vlc_player/
 
 environment:
@@ -22,7 +22,7 @@ dependencies:
   cryptoutils: ^0.4.0
   meta: ^1.2.4
   pedantic: ^1.9.0
-  flutter_vlc_player_platform_interface: ^1.0.3
+  flutter_vlc_player_platform_interface: ^1.0.4
 
 dev_dependencies:
   flutter_test:
