+++ flutter_vlc_player_platform_interface/lib/flutter_vlc_player_platform_interface.dart
@@ -4,6 +4,7 @@ export 'src/enums/hardware_acceleration.dart';
 export 'src/enums/data_source_type.dart';
 export 'src/enums/media_event_type.dart';
 export 'src/enums/renderer_event_type.dart';
+export 'src/enums/subtitles/vlc_subtitle_text_direction.dart';
 
 export 'src/events/media_event.dart';
 export 'src/events/renderer_event.dart';
@@ -14,3 +15,7 @@ export 'src/utils/options/vlc_player_options.dart';
 export 'src/utils/options/vlc_rtp_options.dart';
 export 'src/utils/options/vlc_stream_output_options.dart';
 export 'src/utils/options/vlc_video_options.dart';
+export 'src/utils/options/vlc_subtitle_options.dart';
+
+export 'src/utils/helpers/subtitles/vlc_subtitle_color.dart';
+export 'src/utils/helpers/subtitles/vlc_subtitle_thickness.dart';
