+++ flutter_vlc_player_platform_interface/lib/src/enums/subtitles/vlc_subtitle_text_direction.dart
@@ -1,0 +1,14 @@
+enum VlcSubtitleTextDirection {
+  ltr,
+  rtl,
+  auto,
+}
+
+extension VlcSubtitleTextDirectionExtensionMap on VlcSubtitleTextDirection {
+  static const valueMap = {
+    VlcSubtitleTextDirection.ltr: 0,
+    VlcSubtitleTextDirection.rtl: 1,
+    VlcSubtitleTextDirection.auto: 2,
+  };
+  int get value => valueMap[this];
+}
