+++ flutter_vlc_player_platform_interface/lib/src/utils/options/vlc_player_options.dart
@@ -2,6 +2,7 @@ import 'vlc_advanced_options.dart';
 import 'vlc_audio_options.dart';
 import 'vlc_rtp_options.dart';
 import 'vlc_stream_output_options.dart';
+import 'vlc_subtitle_options.dart';
 import 'vlc_video_options.dart';
 
 class VlcPlayerOptions {
@@ -9,6 +10,7 @@ class VlcPlayerOptions {
     this.advanced,
     this.audio,
     this.video,
+    this.subtitle,
     this.rtp,
     this.sout,
     this.extras,
@@ -17,6 +19,7 @@ class VlcPlayerOptions {
   final VlcAdvancedOptions advanced;
   final VlcAudioOptions audio;
   final VlcVideoOptions video;
+  final VlcSubtitleOptions subtitle;
   final VlcRtpOptions rtp;
   final VlcStreamOutputOptions sout;
   final List<String> extras;
@@ -26,6 +29,7 @@ class VlcPlayerOptions {
     if (advanced != null) options.addAll(advanced.options);
     if (audio != null) options.addAll(audio.options);
     if (video != null) options.addAll(video.options);
+    if (subtitle != null) options.addAll(subtitle.options);
     if (rtp != null) options.addAll(rtp.options);
     if (sout != null) options.addAll(sout.options);
     if (extras != null) options.addAll(extras);
